stripe_worker/requirements.txt
stripe
confluent_kafka
python-dotenv
